Document Skeleton 99% Complete
### Front and Back-end connected
Now React and Express are communicating correctly. This also entails the front end can send and receive data from the database and viceversa since they'll use express as their middleman. Now we just need to create the middleware/routes for express to go to mongo, retrieve data and include that data in the response to the frontend.

### example.env
Easy way to setup env variables when we only need to copy the example.env and change it to setup.env

### gitignore
added the node_modules from client side.