Starting the Map Generator
So albeit at first I had the idea to manually insert the data for the map, I quickly realized that would require for me to insert 324 individual tiles for an 18x18 grid.

Therefore, I decided to have some fun and create a random map generator to have our first test map.

This will involve creating one half of the map, copying that half and then inverting and flipping it and using that as the other half of the map (so we have like a mirror/symmetry).