Randomly generating map
## createMap.js
Can actually create a symmetrical random map. This is just in JSON format however, I still need to code in the frontend for it to take this JSON/input from the backend and render a map correctly. Nonetheless, thats the easy part.

The random map mirrors itself so it should create two sides of the map that are equal.

## Upcoming features
currently all tiles are set at random absolutely. Meaning out of the 7 tiles we have, all have an equal chance of being picked.

Nonetheless! We can improve our current random tile picking process to increase the % of certain tiles being picked and decrease the % of others. For example, we want mountains (a hard, impassable terrain) to be rare while plains and roads (easy to pass terrain) have a higher chance of appearing because that creates a map that is better to play in and not some random monstruosity.