Rendering map through json in React
### Rendering the map
We make an array and push an element with a class "fo" (forest) or other names such as "ci" (city) for each item in our json. Then this div with a class "fo" has a css styling to give it a background of that tile (so the fo class gives a forest background).

As said before, currently I'm trying to mimic a mirror effect, meaning that for each city in one tile, in its opposite/mirror tile, the same city/terrain type exists.

### .gitignore
Added all json files inside of scripts since we don't need to clutter our repo with randomly generated maps.

### Images
Added all the images needed for our background tiles.

### Styling
Added some styles to correctly display our tiles.

### webpack.config.js
Now allows to send the randomMap request to our express backend.