Fix factory, mountain and city not rendering correctly
## Mountains cities render correctly
So now the map actually renders correctly. The biggest issue that made it not render correctly is the fact that mountains and cities are not squares. They have the same width as every other tile (16px) but not the same height (21px).

Therefore, they would not render correctly. The solution was simply making them have the absolute property so they could go over other tiles and give them an increase in sizing.

## Tile sizing
All tiles are 16px * 16px, but some are 16 * 21. Since we don't want to be looking at such squares. FOR NOW, we will be multiplying that by three to have 48 * 48 and 48 * 63 tile sizes.

## Pseudo-Random generation
The tiles are now not entirely random, now the array has its function terrainParameters() which adds tile types to an array that holds 100 values, so when 50 "plains"/"pl" strings are added to the array, there is a 50% chance they will be used everytime we generate a random tile. And so on with the other tiles having smaller %.