Create AWBW map parser and nearest neighbor image rendering
## Image rendering
if we render our images using image-rendering nearest-neighbor, our pixels

## Map Parsing
so currently AWBW maps can be accessed through text files, which represent every tile type in numbers. I had to manually lookup what each number meant and start typing out the type of tile we would parse it to (such as 1 = pl0 which means plains number 0).

The AWBW map parser is not complete but it has enough information to recreate the famous land only map Caustic Finale

## Random != Parsed
I made sure to give the randomMapGenerator.js, its functions routes and so on a different name to specify its random and not just any map