Change tile rendering to 3 letter sequence and @each for rendering
## 3 letter sequence (fo = fo0)
So before our tiles only needed 2 letters to be displayed such as forest being fo. Nonetheless, some tiles like roads, rivers and shoals, have different orientations and curves. To separate those, we now also incorporate a number, so road can be ro1, ro2 and so on. This helps us know which orientation and curve to use for the road.

### image additions
Nothing to explain here, just added a lot of different tile images with their respective folders.

## @each loop
Instead of repeating ourselves a thousand times to type ".pl background: url..." and so on, I've put our backgrounds into an array which gets processed by an @each loop and therefore cuts hundreds of lines we would have had to write but instead making it just a few.