Complete Caustic Finale Render and Tile Cursor
## Map Rendering
Now the map Caustic Finale is fully rendered in our /game page, this is the current map I see as our test map. It's one of the most popular maps and it only involves land units so we don't have to worry about air or sea warfare.

### Map Cursor
Now when hovering over a tile, we have the in-game cursor displaying correctly. It is not perfectly centered but since we are working towards a demo, we first need to get the game running and then worry about polishing it.

## Images
More images were added or refactored for a better naming convention and for them to be rendered correctly.

## TODO
Added some TODO here and there so we can remember at some point to do them.