Calculates movements correctly without taking terrain into account
## Pathfinding
I still need to test it and make it draw the tiles in the front-end but currently it seems to be calculating the correct amount of tiles a unit moving 6 tiles could achieve if all terrain costed just 1 movement to transverse.

## tinyQueue
this dependency was added since making our own queue could be done later but right now we need a functioning game before April 14 so we wont bother implementing our own queue.