Insertion of Unit and Terrain Datasets, changes done to map reading
## Moving to an object based tile set
Currently, we were using an array oriented style to store our gamestate data. However, for the sake of simplicity, I've decided its better to move on with an object oriented one since that lets us add more data to a tile in an easy to read manner instead of creating some monster that goes pl000210os00 that might be a bit faster but harder to work with.

## Data Sets
I added some datasets of units and the cost of movement in certain terrains that I found in https://github.com/kbuzsaki/awbw_enhancements/blob/dev/lib/game_data.js . I added some small tweaks and changes to naming conventions.

## Taking out React Strict Mode
Strict mode was making me sick with its double rendering so I took it out for now.