Upload gamestate to db and render gamestate from db's data
### createGame.js
This file is just a temporary file to test the concepts of creating a game and also getting a game state. These routes will later be modified to take input from the user creating a game or receiving game data.

### Map parsers are now offline
The awbwMapParser.js and randomMapGenerator.js are both now unused. This is because we have parsed thorugh Caustic Finale correctly and we don't need any other maps at the moment. Therefore, they are still very useful but have been put to sleep for now/

### websocket.js
Still doing testing but have stopped because I need to be able to send data to the database, make sure it gets saved and then emit that new data to other clients. So before continuing socket.io development, I must first establish a basic get/post request routes and then use socket.io