Menus, unit building and rendering of all blue moon and orange star units
### Images
Just added all the images for both Orange Star and Blue Moon. Also added the adequate css classes for these to be rendered correctly.

# Unit Building
We can now produce units! By clicking a factory. a small menu will pop up with all the available units.

### Menu
Clicking a unit on the menu which displays all the units will create said unit!

### Country Check
the menu/base also checks which country owns it and displays the colored units of the correct side.

## Parsing Map
Now includes mapData on it, which makes it easier to get the mapName, author and other details without needing for it to be part of the gameState.

It also determines the ownership of a tile by checking to which country it belongs too.

Also added other details to the tank unit.

### Standarization
I continue to work to make sure we don't adequately re-use functions for both units and properties so we don't need the same function twice to render the same menu but just do it one time and check if its an unit or property.