Finished the styling on login and register
The functionality should not be affected at all. Only the styling should have changed. I have not tested if it works but I believe 100% that is does.
Also, I created a small container. This small container is used for only the login and register page currently.
The input field for type "text" is now styled and is located in template.sass

Minor:
- removed primary button styling (this styling is now on button)