Pass Turn, used/unused status, health and capture icon rendering,
### isUsed
Now units have the isUsed attribute. This determines whether a unit has already done an action and is in fact, used. This adds a class to grey them out and makes them unable to move again until the turn is passed.

### Pass turn
Its a bit "simplistic" right now but we have a pass turn button to "free" all used units so they can move again, still needs some work regarding only the "active" player (whoever's turn it is) should be able to move while they pass their turn.

### Health and Capture Icon
Units now display their current health and capture icon correctly! Also, when a unit captures a building, it takes into account their current health! Because to capture we need to put in 20 points, so an infantry with 10 health captures once, thats 10, captures again thats 20. However, an infantry with 5 health needs to capture 4 times to get to 20.