Detect Valid Targets and Attack them
### Detect possible targets and valid targets
Currently we have validTargets.js to find the targets our unit can fire at. This is run from their new position. It also highlights the enemy units in "range" so that the user has a better idea as well

### Attack
Right now, when we are able to attack an unit, we can just click on our desired tile and right click on the target and damage will be applied through the damageCalculator.js function! It still needs some help since any targets to the left of the unit are blocked by the attack menu...

###