Fix login issue and add flash messages
### Login issue
login form wasn't actually sending its data to the post request, now it actually does check if user is or isn't logged in.

### Flash Message
On the previous note, now we can actually inform the user whether they logged in or not via flash message. This will be much better than simply redirecting the user