Fix attack bug and hp display bug
### Attack Bug
The attackAction had no use before. Nonetheless, when we would attack an unit, our menu would be on the way if the enemy unit was to the left of our unit, making it impossible to attack it. Now that the attackAction hides our menu, we can easily attack units blocked by the menu.

### hp display bug
When moving to a new tile, HP wouldn't render on the new tile (so our 5hp infantry would appear as full HP while moving between tiles.)