Income Management Complete and turn handling
### Economy Management
Now the app recognizes how many properties, unit and funding we have. Quite a bit is hardcoded but its not the end of the world. Right now what we need is an MVP, so we can worry about it later. This makes it so everytime we capture a property or pass a turn, the correct funding is added.

### Building units
Now you can only build units if it is your turn AND if you can afford it! Otherwise, the menu will appear with greyed out and strikethrough elements, showing to the user that these options are not available.

### Turn Handling
Now the game actually checks whose turn it is. This means only the active player can perform actions while the inactive player cant. The former has their "player card" highlighted while the latter has a greyed out card.