startgames page is now connected to the database
Added the ability to create a game to startGames collection with the username of the currently logged in user for orangeStar player.
Next: show only current games to user in currentGame page

Created an API endpoint to retrieve currentGames and startGames. StartGames.js in the frontend now has the data from the database from doing an axios get on "/getStartGames".

Placed createGame.js to the bottom of server.js so that we can access the user session in createGame.js.

Created /userInfo API endpoint for the header.

Removed localStorage from LoginForm.js

Formatting webpack with prettier