Users can join games in /startgames page
Clicking on a GameEntry in /startgames page executes an axios.post to /joinGame.
This will get and delete the game from startGame collection and then add it currentGame collection with
the user set to the blueMoon team.
If the user is on orangeStar, the server will not allow them to join the game.

See server.js .post("/joinGame") for more information on joining a game.