Polish Combat Outcome/Battle Forecast and fix unit attacking from any tile.
### Battle Forecast
Has been polished to look cleaner, display correctly and be easier to understand for new players.

### Bugfix: Attack from any tile
Units were able to attack any unit in-range from any tile they could move to. So if an enemy tank was in range of your tank, you could move your tank to any tile and drag your mouse out of the movement tiles and just click on the enemy tank, initiating combat from the last tile hovered. This isn't supposed to happen since you should be adjacent to an unit to be able to attack.