Creating a new game is now done in /newgame
When you press the play button in the /newgame page, it will send an axios post to /createNewGame
with fields:
{
selectedCO: selectedCO,
selectedCountry: selectedCountry,
selectedMap: selectedMap,
};
A new gameDocument will be inserted into startGame collection where the orangeStar.CO will be
selectedCO from above. The blueMoon fields are now empty strings except for username where it is
"..." There is also now a startDate attribute in each game.

Game entries in /startgames will have the blueMoon character be all black and have the username "..."

The /newgame page, /startgames page, and /currentgames page all checks if the user is logged in. If not, it will redirect them to /login. The /newgame page authentication is done with /authenticateUser where it only uses the loggedIn middleware.

Next steps:
SelectCharacter.js, modal.sass, and util.js will be for further implementation of the feature where a user joins another game. In the future, when a user presses on a pending game, the website will prompt them to select a character.