Save game actions in database and apply websockets
### Save actions to DB
Currently everytime an action is made in the game, it is saved to the database. If at some point an action falls through, the user is redirected to the home page so they can know that they are stopped from playing a game that isn't connected to the database anymore.

### Websockets
Websockets have been installed and seem to work correctly, more testing in the AWS EC2 needs to be done to see if they will actually work!