Joining games now requires picking a character
When a user clicks on a game on /startgames, a modal pops up that prompts the user to select a character to play as in the game.
Upon clicking the join button, the game is moved from collection "startGame" to "currentGame".
The following attributes are changed in the game document:
  myDoc.playerState.blueMoon.username = req.session.username;
  myDoc.playerState.blueMoon._id = req.session._id;
  myDoc.playerState.blueMoon.CO = req.body.selectedCO;
(see server.js)

The selected character section from /newgame has been made into it own component, SelectCharacter.js. This component
is used for /newgame and /startgame.

misc:
various sass file changes and additions