WebSocket improvement and other bug fixes
### Websocket
The websocket was getting reconnected to a new session everytime someone logged in. This would mean 1 action would be sent 5 to 30 times instead of just once, and this would be incremental so 5 > 10 > 19 > 27 > 39 and around that just crash the browser, needing to be reloaded to start the whole incremental process from 0 until it crashes again. Hopefully this fixes it.

### Property/Income glitch
Before, if you had build or captured something and refreshed the page, the database wouldnt save the money you spent (so you could buy units, refresh and keep your gold).

Same happened with properties, if you capture and then refresh, it wouldnt update your new income.