Access Current Games, Fix Registration
### Current Games
Now if you click on a current game, you can actually see the specific game by ID.

### Registration
There was a bug making registration difficult, had to manually add ID 0 an 1 to the userbase so it would be fixed.

### Minor Tweaks
A lot of other files have received other very minor tweaks for better performance.