Web socket rooms
### Web socket rooms
Websockets now have rooms so data is only sent to people inside a game rather than everyone on every game (which was causing alot of issues if 2 people were playing different matches).

### User validation
now game actions check if the user matches the turn/id of the active player so only the active player can play.

### Remove Console Logs
removed some console logs here and there so the server is cleaner.

### Stress Test Homepage
going to ask some people in the AWBW community to stress test the application so the homepage has been changed to make it easy for them to register and find a match.