Try catch, My games area start, limit username/password length
### Create/Join game, save that gameID in user
Now when an user joins or creates a game, the ID of that game is saved in an array in their profile. This will help us setup a "my Games" area

### Max length for usernames
Limit username creation to just 14 characters, dont want no clowns tanking the server with an username above hundreds of characters

### Unprotect normal routes
Stuff like currentGames should be unprotected, people should be able to just look at live matches.