Your Games
### Display user's games if they are logged in
Now if an user is checking current games and they are logged in, they will see the games they are involved in at first

### VS Image
Got updated with the same image but much higher quality.

### Limit Amount of games people can make/join
Now people can only make a game if they arent involved in 3 games and they can only join a game if they arent involved in 5 games. This is to avoid people spamming games/joining a bunch of matches they arent going to play.